Fix warnings during generation due to missing summary and description (#178)
Currently, if you run `make generate | grep WARN` there are thousands of
warnings due to endpoints missing summaries or descriptions.

The overwhelming number of warnings makes it harder to spot genuine
errors in the templates - e.g. #177.

This change wraps each unconditional reference to summary or description
in an `{{#if}}` to cure the warnings.