Sync OpenAPI: Unreleased plugin changes
This PR has been generated from a local Vault working copy which has
had:
```
go get github.com/hashicorp/vault-plugin-{secrets-{gcp,kv},auth-jwt}@main
go mod tidy
```
run on it, to pull in some unreleased changes in plugins.

You may or may not want to merge this right now, as it'll mean anyone
else syncing the OpenAPI has to know and keep doing that, to avoid
reverting changes.

I'm happy for it to stand unmerged as a visualization of pending
incoming changes.

There are some further changes it might be worth getting into the
plugins, before a full release cycle is done:

* https://github.com/hashicorp/vault-plugin-secrets-kv/pull/120
* https://github.com/hashicorp/vault-plugin-secrets-gcp/pull/190